% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_count.R
\name{add_count}
\alias{add_count}
\title{add_count}
\usage{
add_count(...)
}
\arguments{
\item{...}{parameters for dplyr::add_count}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for add_count
}
\details{
The wisegroup package provides a wrapper for add_count that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:add_count]{dplyr::add_count()}}.

Use \code{\link[=add_count...]{add_count...()}} to retain groups as per \code{dplyr::add_count}, whilst
signalling this in your code.
}
\seealso{
dplyr::add_count, add_count...
}
\author{
Miles McBain
}
