% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tally.R
\name{add_tally}
\alias{add_tally}
\title{add_tally}
\usage{
add_tally(...)
}
\arguments{
\item{...}{parameters for dplyr::add_tally}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for add_tally
}
\details{
The wisegroup package provides a wrapper for add_tally that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:add_tally]{dplyr::add_tally()}}.

Use \code{\link[=add_tally...]{add_tally...()}} to retain groups as per \code{dplyr::add_tally}, whilst
signalling this in your code.
}
\seealso{
dplyr::add_tally, add_tally...
}
\author{
Miles McBain
}
