% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{count}
\usage{
count(...)
}
\arguments{
\item{...}{parameters for dplyr::count}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for count
}
\details{
The wisegroup package provides a wrapper for count that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:count]{dplyr::count()}}.

Use \code{\link[=count...]{count...()}} to retain groups as per \code{dplyr::count}, whilst
signalling this in your code.
}
\seealso{
dplyr::count, count...
}
\author{
Miles McBain
}
