% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill...}
\alias{fill...}
\title{fill}
\usage{
fill...(...)
}
\arguments{
\item{...}{parameters for tidyr::fill}
}
\value{
an ungrouped dataframe
}
\description{
Grouping wrapper for fill
}
\details{
The wisegroup package provides a wrapper for fill that behaves as per the
original with the added bonus of signalling returned data could be grouped in source code.
This helps avoid mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[tidyr:fill]{tidyr::fill()}}.

Use \code{\link[wisegroup:fill]{wisegroup::fill()}} to automatically remove groups after applying tidyr::fill.
}
\seealso{
tidyr::fill, fill...
}
\author{
Miles McBain
}
