% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{fill}
\usage{
fill(...)
}
\arguments{
\item{...}{parameters for tidyr::fill}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for fill
}
\details{
The wisegroup package provides a wrapper for fill that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [tidyr::fill()].

Use [fill...()] to retain groups as per `tidyr::fill`, whilst
signalling this in your code.
}
\seealso{
tidyr::fill, fill...
}
\author{
Miles McBain
}
