% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate...}
\alias{mutate...}
\title{mutate}
\usage{
mutate...(...)
}
\arguments{
\item{...}{parameters for dplyr::mutate}
}
\value{
an ungrouped dataframe
}
\description{
Grouping wrapper for mutate
}
\details{
The wisegroup package provides a wrapper for mutate that behaves as per the
original with the added bonus of signalling returned data could be grouped in source code.
This helps avoid mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::mutate()].

Use [wisegroup::mutate()] to automatically remove groups after applying dplyr::mutate.
}
\seealso{
dplyr::mutate, mutate...
}
\author{
Miles McBain
}
