% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{mutate}
\usage{
mutate(...)
}
\arguments{
\item{...}{parameters for dplyr::mutate}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for mutate
}
\details{
The wisegroup package provides a wrapper for mutate that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:mutate]{dplyr::mutate()}}.

Use \code{\link[=mutate...]{mutate...()}} to retain groups as per \code{dplyr::mutate}, whilst
signalling this in your code.
}
\seealso{
dplyr::mutate, mutate...
}
\author{
Miles McBain
}
