% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_all.R
\name{mutate_all}
\alias{mutate_all}
\title{mutate_all}
\usage{
mutate_all(...)
}
\arguments{
\item{...}{parameters for dplyr::mutate_all}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for mutate_all
}
\details{
The wisegroup package provides a wrapper for mutate_all that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::mutate_all()].

Use [mutate_all...()] to retain groups as per `dplyr::mutate_all`, whilst
signalling this in your code.
}
\seealso{
dplyr::mutate_all, mutate_all...
}
\author{
Miles McBain
}
