% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_at.R
\name{mutate_at}
\alias{mutate_at}
\title{mutate_at}
\usage{
mutate_at(...)
}
\arguments{
\item{...}{parameters for dplyr::mutate_at}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for mutate_at
}
\details{
The wisegroup package provides a wrapper for mutate_at that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::mutate_at()].

Use [mutate_at...()] to retain groups as per `dplyr::mutate_at`, whilst
signalling this in your code.
}
\seealso{
dplyr::mutate_at, mutate_at...
}
\author{
Miles McBain
}
