% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_if.R
\name{mutate_if}
\alias{mutate_if}
\title{mutate_if}
\usage{
mutate_if(...)
}
\arguments{
\item{...}{parameters for dplyr::mutate_if}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for mutate_if
}
\details{
The wisegroup package provides a wrapper for mutate_if that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::mutate_if()].

Use [mutate_if...()] to retain groups as per `dplyr::mutate_if`, whilst
signalling this in your code.
}
\seealso{
dplyr::mutate_if, mutate_if...
}
\author{
Miles McBain
}
