% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frac.R
\name{sample_frac}
\alias{sample_frac}
\title{sample_frac}
\usage{
sample_frac(...)
}
\arguments{
\item{...}{parameters for dplyr::sample_frac}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for sample_frac
}
\details{
The wisegroup package provides a wrapper for sample_frac that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:sample_frac]{dplyr::sample_frac()}}.

Use \code{\link[=sample_frac...]{sample_frac...()}} to retain groups as per \code{dplyr::sample_frac}, whilst
signalling this in your code.
}
\seealso{
dplyr::sample_frac, sample_frac...
}
\author{
Miles McBain
}
