% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n.R
\name{sample_n}
\alias{sample_n}
\title{sample_n}
\usage{
sample_n(...)
}
\arguments{
\item{...}{parameters for dplyr::sample_n}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for sample_n
}
\details{
The wisegroup package provides a wrapper for sample_n that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:sample_n]{dplyr::sample_n()}}.

Use \code{\link[=sample_n...]{sample_n...()}} to retain groups as per \code{dplyr::sample_n}, whilst
signalling this in your code.
}
\seealso{
dplyr::sample_n, sample_n...
}
\author{
Miles McBain
}
