% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{slice}
\usage{
slice(...)
}
\arguments{
\item{...}{parameters for dplyr::slice}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for slice
}
\details{
The wisegroup package provides a wrapper for slice that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:slice]{dplyr::slice()}}.

Use \code{\link[=slice...]{slice...()}} to retain groups as per \code{dplyr::slice}, whilst
signalling this in your code.
}
\seealso{
dplyr::slice, slice...
}
\author{
Miles McBain
}
