% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\title{summarise}
\usage{
summarise(...)
}
\arguments{
\item{...}{parameters for dplyr::summarise}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarise
}
\details{
The wisegroup package provides a wrapper for summarise that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:summarise]{dplyr::summarise()}}.

Use \code{\link[=summarise...]{summarise...()}} to retain groups as per \code{dplyr::summarise}, whilst
signalling this in your code.
}
\seealso{
dplyr::summarise, summarise...
}
\author{
Miles McBain
}
