% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_all.R
\name{summarise_all}
\alias{summarise_all}
\title{summarise_all}
\usage{
summarise_all(...)
}
\arguments{
\item{...}{parameters for dplyr::summarise_all}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarise_all
}
\details{
The wisegroup package provides a wrapper for summarise_all that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::summarise_all()].

Use [summarise_all...()] to retain groups as per `dplyr::summarise_all`, whilst
signalling this in your code.
}
\seealso{
dplyr::summarise_all, summarise_all...
}
\author{
Miles McBain
}
