% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_at.R
\name{summarise_at}
\alias{summarise_at}
\title{summarise_at}
\usage{
summarise_at(...)
}
\arguments{
\item{...}{parameters for dplyr::summarise_at}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarise_at
}
\details{
The wisegroup package provides a wrapper for summarise_at that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::summarise_at()].

Use [summarise_at...()] to retain groups as per `dplyr::summarise_at`, whilst
signalling this in your code.
}
\seealso{
dplyr::summarise_at, summarise_at...
}
\author{
Miles McBain
}
