% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_if.R
\name{summarise_if}
\alias{summarise_if}
\title{summarise_if}
\usage{
summarise_if(...)
}
\arguments{
\item{...}{parameters for dplyr::summarise_if}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarise_if
}
\details{
The wisegroup package provides a wrapper for summarise_if that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:summarise_if]{dplyr::summarise_if()}}.

Use \code{\link[=summarise_if...]{summarise_if...()}} to retain groups as per \code{dplyr::summarise_if}, whilst
signalling this in your code.
}
\seealso{
dplyr::summarise_if, summarise_if...
}
\author{
Miles McBain
}
