% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all.R
\name{summarize_all}
\alias{summarize_all}
\title{summarize_all}
\usage{
summarize_all(...)
}
\arguments{
\item{...}{parameters for dplyr::summarize_all}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarize_all
}
\details{
The wisegroup package provides a wrapper for summarize_all that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see [dplyr::summarize_all()].

Use [summarize_all...()] to retain groups as per `dplyr::summarize_all`, whilst
signalling this in your code.
}
\seealso{
dplyr::summarize_all, summarize_all...
}
\author{
Miles McBain
}
