% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_if.R
\name{summarize_if}
\alias{summarize_if}
\title{summarize_if}
\usage{
summarize_if(...)
}
\arguments{
\item{...}{parameters for dplyr::summarize_if}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for summarize_if
}
\details{
The wisegroup package provides a wrapper for summarize_if that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:summarize_if]{dplyr::summarize_if()}}.

Use \code{\link[=summarize_if...]{summarize_if...()}} to retain groups as per \code{dplyr::summarize_if}, whilst
signalling this in your code.
}
\seealso{
dplyr::summarize_if, summarize_if...
}
\author{
Miles McBain
}
