% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally}
\alias{tally}
\title{tally}
\usage{
tally(...)
}
\arguments{
\item{...}{parameters for dplyr::tally}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for tally
}
\details{
The wisegroup package provides a wrapper for tally that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:tally]{dplyr::tally()}}.

Use \code{\link[=tally...]{tally...()}} to retain groups as per \code{dplyr::tally}, whilst
signalling this in your code.
}
\seealso{
dplyr::tally, tally...
}
\author{
Miles McBain
}
