% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_all.R
\name{transmute_all}
\alias{transmute_all}
\title{transmute_all}
\usage{
transmute_all(...)
}
\arguments{
\item{...}{parameters for dplyr::transmute_all}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for transmute_all
}
\details{
The wisegroup package provides a wrapper for transmute_all that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:transmute_all]{dplyr::transmute_all()}}.

Use \code{\link[=transmute_all...]{transmute_all...()}} to retain groups as per \code{dplyr::transmute_all}, whilst
signalling this in your code.
}
\seealso{
dplyr::transmute_all, transmute_all...
}
\author{
Miles McBain
}
