% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_at.R
\name{transmute_at}
\alias{transmute_at}
\title{transmute_at}
\usage{
transmute_at(...)
}
\arguments{
\item{...}{parameters for dplyr::transmute_at}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for transmute_at
}
\details{
The wisegroup package provides a wrapper for transmute_at that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:transmute_at]{dplyr::transmute_at()}}.

Use \code{\link[=transmute_at...]{transmute_at...()}} to retain groups as per \code{dplyr::transmute_at}, whilst
signalling this in your code.
}
\seealso{
dplyr::transmute_at, transmute_at...
}
\author{
Miles McBain
}
