% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_if.R
\name{transmute_if}
\alias{transmute_if}
\title{transmute_if}
\usage{
transmute_if(...)
}
\arguments{
\item{...}{parameters for dplyr::transmute_if}
}
\value{
an ungrouped dataframe
}
\description{
Ungrouping wrapper for transmute_if
}
\details{
The wisegroup package provides a wrapper for transmute_if that always returns
ungrouped data. This avoids mistakes associated with forgetting to call ungroup().

For original documentation see \code{\link[dplyr:transmute_if]{dplyr::transmute_if()}}.

Use \code{\link[=transmute_if...]{transmute_if...()}} to retain groups as per \code{dplyr::transmute_if}, whilst
signalling this in your code.
}
\seealso{
dplyr::transmute_if, transmute_if...
}
\author{
Miles McBain
}
